feat: enhance dashboard with ChartsAnalysisSection and refactor StatCard styles. Pending: take to ssr dashboard page
