+++ src/frontend/factchecker-front/src/components/ui/dashboard/index.ts
@@ -1,1 +1,2 @@
 export { StatCardsGrid } from './StatCardsGrid';
+export { ChartsAnalysisSection } from './ChartsAnalysisSection';
